Json input and graphs
Modified the input of the ANN parameters from a .txt file to a .json, which made it much easier to understand and use.
Also added an accuracy bar and a target chart to greatly improve the visualization and undestanding of the results.

Next step is to include a "improvement rate" graph to show how fast the IA is improving during training in real time.